Clave de Hashing ahora es unsigned long by Lucas chico 10. Estado: Daniel  esta con memory leaks y luego con informes. MartinH esta con conteo e informes. Lucas   esta con la interfaz del administrador. A demas falta permitirle bajas y modificaciones. MartinQ esta con la persistencia del arbol. Ignacio esta con el balanceo del arbol.

+++ trunk/src/Archivos/Hashing/Registro.cpp
@@ -53,7 +53,7 @@ Entidad *Registro::getContenido(){
 	return entidad;
 }
 
-int Registro::obtenerClave(){
+unsigned long Registro::obtenerClave(){
 	return this->clave;
 }
